Skygrab is now a tab.
If it is selected repeatedly on the same spectrum, it will not load more than once, unless you switch back and forth between spectra. Ask Fynbo if it should be a window or a tab.

ssPicture and nav manipulate a new boolean, skygrabNotLoaded, in the model. Should this be refactored?